+++ src/stellar_ram.c
@@ -1,0 +1,30 @@
+/*******************************************************************************
+ *   Ledger Stellar App
+ *   (c) 2017-2018 Ledger
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ ********************************************************************************/
+
+#include "os.h"
+#include "cx.h"
+#include "stellar_types.h"
+#include "stellar_api.h"
+#include "stellar_vars.h"
+
+#include "os_io_seproxyhal.h"
+#include "u2f_service.h"
+
+stellar_context_t ctx;
+ux_state_t ux;
+unsigned char G_io_seproxyhal_spi_buffer[IO_SEPROXYHAL_BUFFER_SIZE_B];
+volatile u2f_service_t u2fService;
